This takes care of several compiler warnings and brings py-ogg2 to sync with libogg2, specifically, rewritting OggPage's setattr functions.
There seems to be a problem with OggSyncState that I have yet to 
resolve.  I hope it's in py-ogg2 and not libogg2.  Apparently, the Sync 
buffer believes that fewer bytes have been submitted than have actually 
been.  This results, in my test scripts, in files being outputted that 
are half the size as they should be.